Refactor folders and remove unnecessary files
fix #14
