--- DD2480-CI/.idea/workspace.xml
@@ -1,139 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="AutoImportSettings">
-    <option name="autoReloadType" value="SELECTIVE" />
-  </component>
-  <component name="ChangeListManager">
-    <list default="true" id="badd97bb-120e-4cef-a476-0e558e54772a" name="Changes" comment="">
-      <change afterPath="$PROJECT_DIR$/src/test/java/mavenBuilderTest/MavenInvokerBuilderTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../.gitignore" beforeDir="false" afterPath="$PROJECT_DIR$/../.gitignore" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/.gitignore" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/encodings.xml" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/misc.xml" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/vcs.xml" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/pom.xml" beforeDir="false" afterPath="$PROJECT_DIR$/pom.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/mavenBuilder/MavenInvokerBuilder.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/mavenBuilder/MavenInvokerBuilder.java" afterDir="false" />
-    </list>
-    <option name="SHOW_DIALOG" value="false" />
-    <option name="HIGHLIGHT_CONFLICTS" value="true" />
-    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
-    <option name="LAST_RESOLUTION" value="IGNORE" />
-  </component>
-  <component name="FileTemplateManagerImpl">
-    <option name="RECENT_TEMPLATES">
-      <list>
-        <option value="Class" />
-      </list>
-    </option>
-  </component>
-  <component name="Git.Settings">
-    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$/.." />
-  </component>
-  <component name="ProjectColorInfo">{
-  &quot;associatedIndex&quot;: 5
-}</component>
-  <component name="ProjectId" id="2c5fqKJzK9HDQ4DqWyAOoqiwj3w" />
-  <component name="ProjectViewState">
-    <option name="hideEmptyMiddlePackages" value="true" />
-    <option name="showLibraryContents" value="true" />
-  </component>
-  <component name="PropertiesComponent"><![CDATA[{
-  "keyToString": {
-    "JUnit.MavenInvokerBuilderTest (1).executor": "Run",
-    "JUnit.MavenInvokerBuilderTest.build.executor": "Run",
-    "JUnit.MavenInvokerBuilderTest.executor": "Run",
-    "RunOnceActivity.OpenProjectViewOnStart": "true",
-    "RunOnceActivity.ShowReadmeOnStart": "true",
-    "git-widget-placeholder": "issue/8",
-    "kotlin-language-version-configured": "true",
-    "last_opened_file_path": "/Users/liangtianning/Documents/GitHub/dd2480-CI/DD2480-CI",
-    "node.js.detected.package.eslint": "true",
-    "node.js.detected.package.tslint": "true",
-    "node.js.selected.package.eslint": "(autodetect)",
-    "node.js.selected.package.tslint": "(autodetect)",
-    "nodejs_package_manager_path": "npm",
-    "vue.rearranger.settings.migration": "true"
-  }
-}]]></component>
-  <component name="RecentsManager">
-    <key name="MoveClassesOrPackagesDialog.RECENTS_KEY">
-      <recent name="mavenBuilder" />
-    </key>
-  </component>
-  <component name="RunManager" selected="JUnit.MavenInvokerBuilderTest.build">
-    <configuration name="MavenInvokerBuilderTest" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
-      <module name="DD2480-CI" />
-      <extension name="coverage">
-        <pattern>
-          <option name="PATTERN" value="mavenBuilderTest.*" />
-          <option name="ENABLED" value="true" />
-        </pattern>
-      </extension>
-      <option name="PACKAGE_NAME" value="mavenBuilderTest" />
-      <option name="MAIN_CLASS_NAME" value="mavenBuilderTest.MavenInvokerBuilderTest" />
-      <option name="METHOD_NAME" value="" />
-      <option name="TEST_OBJECT" value="class" />
-      <method v="2">
-        <option name="Make" enabled="true" />
-      </method>
-    </configuration>
-    <configuration name="MavenInvokerBuilderTest.build" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
-      <module name="DD2480-CI" />
-      <extension name="coverage">
-        <pattern>
-          <option name="PATTERN" value="mavenBuilder.*" />
-          <option name="ENABLED" value="true" />
-        </pattern>
-      </extension>
-      <option name="PACKAGE_NAME" value="mavenBuilderTest" />
-      <option name="MAIN_CLASS_NAME" value="mavenBuilderTest.MavenInvokerBuilderTest" />
-      <option name="METHOD_NAME" value="build" />
-      <option name="TEST_OBJECT" value="method" />
-      <envs>
-        <env name="MAVEN_HOME" value="$USER_HOME$/Downloads/apache-maven-3.9.6" />
-      </envs>
-      <method v="2">
-        <option name="Make" enabled="true" />
-      </method>
-    </configuration>
-    <list>
-      <item itemvalue="JUnit.MavenInvokerBuilderTest" />
-      <item itemvalue="JUnit.MavenInvokerBuilderTest.build" />
-    </list>
-    <recent_temporary>
-      <list>
-        <item itemvalue="JUnit.MavenInvokerBuilderTest.build" />
-        <item itemvalue="JUnit.MavenInvokerBuilderTest" />
-        <item itemvalue="JUnit.MavenInvokerBuilderTest" />
-      </list>
-    </recent_temporary>
-  </component>
-  <component name="SharedIndexes">
-    <attachedChunks>
-      <set>
-        <option value="jdk-21.0.2-openjdk-21.0.2-4caba194b151-322ed176" />
-        <option value="jdk-21.0.2-openjdk-21.0.2-f644763e9732-a81d5c7c" />
-      </set>
-    </attachedChunks>
-  </component>
-  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
-  <component name="TaskManager">
-    <task active="true" id="Default" summary="Default task">
-      <changelist id="badd97bb-120e-4cef-a476-0e558e54772a" name="Changes" comment="" />
-      <created>1707406549191</created>
-      <option name="number" value="Default" />
-      <option name="presentableId" value="Default" />
-      <updated>1707406549191</updated>
-      <workItem from="1707406550269" duration="1180000" />
-      <workItem from="1707407758319" duration="743000" />
-      <workItem from="1707410615305" duration="456000" />
-      <workItem from="1707411102248" duration="2139000" />
-      <workItem from="1707413599192" duration="279000" />
-      <workItem from="1707413883164" duration="1365000" />
-    </task>
-    <servers />
-  </component>
-  <component name="TypeScriptGeneratedFilesManager">
-    <option name="version" value="3" />
-  </component>
-</project>
