+++ DD2480-CI/src/main/java/ciServer/BuildHistoryManager.java
@@ -1,4 +1,4 @@
-package skeleton;
+package ciServer;
 
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
