+++ DD2480-CI/src/main/java/ciServer/ContinuousIntegrationServer.java
@@ -1,10 +1,8 @@
-package skeleton;
+package ciServer;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.ServletException;
 import java.io.IOException;
-import java.time.LocalDateTime;
-import java.util.List;
 
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.Request;
