+++ DD2480-CI/src/main/java/ciServer/GitRepo.java
@@ -1,4 +1,4 @@
-package skeleton;
+package ciServer;
 
 import java.io.File;
