+++ DD2480-CI/src/main/java/ciServer/GitStatusUpdate.java
@@ -1,4 +1,4 @@
-package skeleton;
+package ciServer;
 
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.HttpPost;
