+++ DD2480-CI/src/main/java/ciServer/RequestParser.java
@@ -1,4 +1,4 @@
-package skeleton;
+package ciServer;
 
 import javax.servlet.http.HttpServletRequest;
 import org.json.simple.JSONObject;
