+++ DD2480-CI/src/test/java/ciServer/BuildHistoryManagerTest.java
@@ -1,4 +1,4 @@
-package skeleton;
+package ciServer;
 
 import org.junit.Test;
 import java.io.File;
