+++ DD2480-CI/src/test/java/ciServer/GitRepoTest.java
@@ -1,4 +1,4 @@
-package skeleton;
+package ciServer;
 
 import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
