+++ DD2480-CI/src/test/java/ciServer/GitStatusUpdateTest.java
@@ -1,4 +1,4 @@
-package skeleton;
+package ciServer;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
