+++ DD2480-CI/src/test/java/ciServer/MavenInvokerBuilderTest.java
@@ -2,11 +2,8 @@ package mavenBuilderTest;
 
 import mavenBuilder.MavenInvokerBuilder;
 import org.apache.maven.shared.invoker.MavenInvocationException;
-import org.eclipse.jgit.api.errors.GitAPIException;
 import org.junit.Test;
-import skeleton.GitRepo;
 
-import java.io.File;
 import java.nio.file.Path;
 import java.nio.file.Paths;
