+++ DD2480-CI/src/test/java/ciServer/RequestParserTest.java
@@ -1,4 +1,4 @@
-package skeleton;
+package ciServer;
 
 import javax.servlet.http.HttpServletRequest;
 import static org.mockito.Mockito.*;
